#28: Fix `shorten-64-to-32`

+++ src/api/hardware/v0.5/hash/hca_sha_miscellaneous.c
@@ -46,7 +46,7 @@
 #include <api/hardware/v0.5/sifive_hca-0.5.x.h>
 
 int32_t hca_sha_block(const metal_scl_t *const scl, hash_mode_t hash_mode,
-                      uint32_t NbBlocks512, const uint8_t *const data_in)
+                      size_t NbBlocks512, const uint8_t *const data_in)
 {
 #if __riscv_xlen == 64
     const uint64_t *in64 = (const uint64_t *)data_in;
